+++ src/HomotechsualBot/Models/SingleMessageChannelConfig.cs
@@ -1,7 +1,0 @@
-namespace DiscordBot.Models;
-
-public class SingleMessageChannelConfig
-{
-    public ulong ChannelId { get; set; }
-    public bool ScanHistoryOnEnable { get; set; } = false;
-}